Doc comments are added
Was working with that great Navigation Bar, but got surprised when I have found no Documentation Comments for its properties, so here they are, feel free to review and make sure the written comments describe the intended behavior and recheck the property `letIndexChange` because I'm not sure what actually it does. my greetings.